Fix system initialization
